unit testing

+++ backend/db/models/questions.py
@@ -5,7 +5,7 @@ from sqlalchemy.orm import relationship
 #it is inherited from Base which is an as_declarative -> makes it possible to create tables automatically
 class Question(Base):
     __tablename__ = "questions" #name in db
-    id = Column(BigInteger, primary_key=True)
+    id = Column(Integer, primary_key=True)
     description = Column(String, nullable=False)
     date_posted = Column(DateTime, nullable=False)
     is_solved = Column(Boolean, default=False)
